[type:e2e] refactor e2e
