+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-common/src/main/java/org/apache/shenyu/e2e/testcase/common/ResourceDataTemplate.java
