+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-common/src/main/java/org/apache/shenyu/e2e/testcase/common/utils/ShenYuSpecUtils.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shenyu.e2e.testcase.common.utils;
+
+import org.apache.shenyu.e2e.testcase.common.specification.ShenYuAfterEachSpec;
+import org.apache.shenyu.e2e.testcase.common.specification.ShenYuBeforeEachSpec;
+import org.apache.shenyu.e2e.testcase.common.specification.ShenYuCaseSpec;
+import org.apache.shenyu.e2e.testcase.common.specification.ShenYuScenarioSpec;
+
+public class ShenYuSpecUtils {
+    
+    /**
+     * build e2e spec.
+     *
+     * @param name name
+     * @param beforeEachSpec before each
+     * @param caseSpec case
+     * @param afterEachSpec after each
+     * @return ShenYuScenarioSpec
+     */
+    public static ShenYuScenarioSpec buildShenYuScenarioSpec(final String name,
+                                                       final ShenYuBeforeEachSpec beforeEachSpec,
+                                                       final ShenYuCaseSpec caseSpec,
+                                                       final ShenYuAfterEachSpec afterEachSpec) {
+        return ShenYuScenarioSpec.builder()
+                .name(name)
+                .beforeEachSpec(beforeEachSpec)
+                .caseSpec(caseSpec)
+                .afterEachSpec(afterEachSpec)
+                .build();
+    }
+}
