+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-divide/src/test/java/org/apache/shenyu/e2e/testcase/divide/plugin/PluginsTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shenyu.e2e.testcase.plugin;
+package org.apache.shenyu.e2e.testcase.divide.plugin;
 
 import org.apache.shenyu.e2e.client.admin.AdminClient;
 import org.apache.shenyu.e2e.client.admin.model.ResourcesData;
