+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-divide/src/test/resources/docker-compose.mysql.yml
